data_preprocessing_research/requirements.txt
# Core AutoML and ML libraries
autogluon>=1.3.0
scikit-learn>=1.6.0
pandas>=2.2.0
numpy>=2.1.0

# Text processing and embeddings
sentence-transformers>=3.0.0
transformers>=4.40.0
tokenizers>=0.19.0

# Data preprocessing and augmentation
imbalanced-learn>=0.12.0  # SMOTE and other sampling techniques
cleanlab>=2.6.0           # Noisy data detection and cleaning
pyod>=2.0.0              # Outlier detection
dedupe>=2.0.0            # Duplicate detection

# Visualization and analysis
matplotlib>=3.10.0
seaborn>=0.13.0
plotly>=5.20.0
wordcloud>=1.9.0

# Experiment tracking and reproducibility
mlflow>=2.10.0
wandb>=0.16.0
hydra-core>=1.3.0        # Configuration management
omegaconf>=2.3.0

# Data handling
datasets>=4.0.0          # HuggingFace datasets
opendatasets>=0.1.0      # Kaggle and other data sources

# Utilities
tqdm>=4.66.0
joblib>=1.4.0
pyyaml>=6.0.0
click>=8.1.0             # CLI interface

# Development and testing
pytest>=8.0.0
pytest-cov>=4.0.0
black>=24.0.0            # Code formatting
isort>=5.13.0            # Import sorting
flake8>=7.0.0            # Linting

# Jupyter and analysis
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Optional: GPU support
# torch>=2.2.0
# torchvision>=0.17.0
